+++ wpull/iostream.py
@@ -9,6 +9,7 @@ import os
 import re
 import socket
 import ssl
+
 import tornado.gen
 import tornado.ioloop
 from tornado.iostream import StreamClosedError