+++ wpull/options.py
@@ -6,6 +6,7 @@ import os
 import ssl
 import sys
 
+import wpull.string
 import wpull.version
 
 
@@ -146,7 +147,7 @@ class AppArgumentParser(argparse.ArgumentParser):
         _logger.debug('Encoding: {0}'.format(encoding))
 
         args = super().parse_args(
-            args=wpull.util.to_str(args, encoding=encoding),
+            args=wpull.string.to_str(args, encoding=encoding),
             namespace=namespace
         )
 
@@ -157,8 +158,8 @@ class AppArgumentParser(argparse.ArgumentParser):
     def get_argv_encoding(cls, argv):
         encoding = 'utf-8'
         stripped_argv = [
-            wpull.util.printable_bytes(arg) for arg in
-            wpull.util.to_bytes(argv, encoding='ascii', error='replace')
+            wpull.string.printable_bytes(arg) for arg in
+            wpull.string.to_bytes(argv, encoding='ascii', error='replace')
         ]
 
         try:
@@ -168,7 +169,7 @@ class AppArgumentParser(argparse.ArgumentParser):
         else:
             encoding = stripped_argv[index + 1]
 
-        return wpull.util.to_str(encoding)
+        return wpull.string.to_str(encoding)
 
     def exit(self, status=0, message=None):
         if self._real_exit: